requirements.txt
numpy
pandas
pyarrow
scipy
scikit-learn
nltk
faiss-cpu
matplotlib
wordcloud
streamlit>=1.39.0
joblib
kaggle
huggingface_hub
transformers
sentence-transformers
gensim
tqdm
ipywidgets
spotipy